js/datachannel_test_2peer/datachanneldict/010.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - DataChannelDict Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
				
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
            var label = "test";
            var dataChannelOptions = {
	            ordered : false,
	            id : 144,
	            maxRetransmits : 333
	        };
            var test = async_test("Set up a DataChannel with the attributes ordered = " + dataChannelOptions.ordered + ", id = " + dataChannelOptions.id + " and maxRetransmits = " + dataChannelOptions.maxRetransmits);
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label, dataChannelOptions);
                    dataChannel.onopen = test.step_func(function() {
                        assert_equals(dataChannel.ordered, dataChannelOptions.ordered, "Can't set ordered: ");
				        assert_equals(dataChannel.id, dataChannelOptions.id, "Can't set id: ");
				        assert_equals(dataChannel.maxRetransmits, dataChannelOptions.maxRetransmits, "Can't set maxRetransmits: ");
                        test.done();
                    });
                });
            } else {
                peerConnection.ondatachannel = test.step_func(function(e) {
                    dataChannel = e.channel;
                    dataChannel.onopen = test.step_func(function() {
                        assert_equals(dataChannel.ordered, dataChannelOptions.ordered, "Can't set ordered: ");
				        assert_equals(dataChannel.id, dataChannelOptions.id, "Can't set id: ");
				        assert_equals(dataChannel.maxRetransmits, dataChannelOptions.maxRetransmits, "Can't set maxRetransmits: ");
                        test.done();
                    });
                });
            }
		</script>
	</body>
</html>
